+++ apps/fluent-tester/src/FluentTester/TestComponents/ButtonExperimental/ButtonSizeTestSection.tsx
@@ -57,15 +57,19 @@ export const ButtonSizeTest: React.FunctionComponent = () => {
       <CompoundButton secondaryContent="Large compound button" size="large" style={commonTestStyles.vmargin}>
         Compound Button
       </CompoundButton>
-      <CompoundButton icon={{ svgSource: svgProps }} secondaryContent="SecondaryContent" size="small" style={commonTestStyles.vmargin}>
-        Content
-      </CompoundButton>
-      <CompoundButton icon={{ svgSource: svgProps }} secondaryContent="SecondaryContent" size="medium" style={commonTestStyles.vmargin}>
-        Content
-      </CompoundButton>
-      <CompoundButton icon={{ svgSource: svgProps }} secondaryContent="SecondaryContent" size="large" style={commonTestStyles.vmargin}>
-        Content
-      </CompoundButton>
+      {svgIconsEnabled && (
+        <>
+          <CompoundButton icon={{ svgSource: svgProps }} secondaryContent="SecondaryContent" size="small" style={commonTestStyles.vmargin}>
+            Content
+          </CompoundButton>
+          <CompoundButton icon={{ svgSource: svgProps }} secondaryContent="SecondaryContent" size="medium" style={commonTestStyles.vmargin}>
+            Content
+          </CompoundButton>
+          <CompoundButton icon={{ svgSource: svgProps }} secondaryContent="SecondaryContent" size="large" style={commonTestStyles.vmargin}>
+            Content
+          </CompoundButton>
+        </>
+      )}
     </View>
   );
 };