+++ apps/fluent-tester/src/FluentTester/TestComponents/ButtonExperimental/E2EButtonTest.tsx
@@ -23,13 +23,12 @@ export const E2EButtonExperimentalTest: React.FunctionComponent = () => {
   return (
     <View>
       <Stack style={stackStyle}>
-        <Button
-          content="This is a button for E2E testing"
-          testID={BUTTONEXPERIMENTAL_TEST_COMPONENT}
-          onClick={onClick}
-          accessibilityLabel={BUTTONEXPERIMENTAL_ACCESSIBILITY_LABEL}
-        />
-        <Button content={BUTTONEXPERIMENTAL_TEST_COMPONENT_LABEL} testID={BUTTONEXPERIMENTAL_NO_A11Y_LABEL_COMPONENT} onClick={onClick} />
+        <Button testID={BUTTONEXPERIMENTAL_TEST_COMPONENT} onClick={onClick} accessibilityLabel={BUTTONEXPERIMENTAL_ACCESSIBILITY_LABEL}>
+          This is a button for E2E testing
+        </Button>
+        <Button testID={BUTTONEXPERIMENTAL_NO_A11Y_LABEL_COMPONENT} onClick={onClick}>
+          {BUTTONEXPERIMENTAL_TEST_COMPONENT_LABEL}
+        </Button>
         {buttonPressed ? <Text testID={BUTTONEXPERIMENTAL_ON_PRESS}>Button Pressed</Text> : null}
       </Stack>
     </View>