+++ apps/fluent-tester/src/FluentTester/TestComponents/MenuButtonExperimental/consts.ts
@@ -1,2 +1,10 @@
-export const HOMEPAGE_MENU_BUTTON_EXPERIMENTAL = 'Homepage_MenuButtonExperimental';
+export const HOMEPAGE_EXPERIMENTAL_MENU_BUTTON = 'Homepage_MenuButtonExperimental';
 export const MENU_BUTTON_EXPERIMENTAL_TESTPAGE = 'MenuButtonExperimental_TestPage';
+
+/* E2E Testing Experimental Menu_Button 1 */
+export const EXPERIMENTAL_MENU_BUTTON_TEST_COMPONENT = 'Experimental_Menu_Button_Test_Component'; // A component on each specific test page
+export const EXPERIMENTAL_MENU_BUTTON_ACCESSIBILITY_LABEL = 'E2E testing Experimental_Menu_Button accessibility label';
+
+/* E2E Testing Experimental Menu_Button 2 */
+export const EXPERIMENTAL_MENU_BUTTON_NO_A11Y_LABEL_COMPONENT = 'Experimental_Menu_Button_No_A11y_label_Component';
+export const EXPERIMENTAL_MENU_BUTTON_TEST_COMPONENT_LABEL = 'Test Experimental_Menu_Button2 - No Accessibility Label'; // A component on each specific test page