+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Mon, 20 Dec 2021 22:56:01 GMT and should not be manually modified.
+This log was last generated on Tue, 21 Dec 2021 22:47:33 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.50.12
+
+Tue, 21 Dec 2021 22:47:33 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-menu-button to v0.2.12
+- Bump @fluentui-react-native/experimental-shimmer to v0.6.8
+- Bump @fluentui/react-native to v0.27.11
+
+## 0.50.11
+
+Tue, 21 Dec 2021 20:56:31 GMT
+
+### Patches
+
+- Remove content prop (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/button to v0.21.8
+- Bump @fluentui-react-native/experimental-activity-indicator to v0.5.6
+- Bump @fluentui-react-native/experimental-avatar to v0.13.8
+- Bump @fluentui-react-native/experimental-button to v0.11.9
+- Bump @fluentui-react-native/experimental-checkbox to v0.4.8
+- Bump @fluentui-react-native/experimental-expander to v0.3.7
+- Bump @fluentui-react-native/experimental-menu-button to v0.2.11
+- Bump @fluentui-react-native/experimental-shimmer to v0.6.7
+- Bump @fluentui-react-native/experimental-tabs to v0.4.8
+- Bump @fluentui-react-native/experimental-text to v0.7.7
+- Bump @fluentui-react-native/framework to v0.7.7
+- Bump @fluentui-react-native/icon to v0.11.1
+- Bump @fluentui-react-native/interactive-hooks to v0.11.8
+- Bump @fluentui-react-native/stack to v0.6.7
+- Bump @fluentui/react-native to v0.27.10
 
 ## 0.50.10