+++ packages/experimental/MenuButton/CHANGELOG.md
@@ -1,8 +1,27 @@
 # Change Log - @fluentui-react-native/experimental-menu-button
 
-This log was last generated on Mon, 20 Dec 2021 22:56:00 GMT and should not be manually modified.
+This log was last generated on Tue, 21 Dec 2021 22:47:33 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.2.12
+
+Tue, 21 Dec 2021 22:47:33 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.14.5
+
+## 0.2.11
+
+Tue, 21 Dec 2021 20:56:31 GMT
+
+### Patches
+
+- Remove content prop (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/contextual-menu to v0.14.4
+- Bump @fluentui-react-native/experimental-button to v0.11.9
+- Bump @fluentui-react-native/framework to v0.7.7
 
 ## 0.2.10